Improve training (see description)
1. Use all training data for cross validation
2. Prioritize bigrams in compromised k frequencies

E.g. k = 3, but if we get an empty count, we check for k = 2, ...
Until we max out the compromise (until k = 2 in this case)

If still empty, we get only the bigrams of the over-compromised k
(i.e. k = 1). If the counter remains empty, we will then resort to
unigrams.

Signed-off-by: Jia Yee <jiayeerawr@gmail.com>